pyproject.toml
[tool.poetry]
name = "art-of-war"
version = "0.1.0"
description = "a chatbot to interact with the Art of War"
authors = ["Saul Thompson saulthomp@outlook.com"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
psycopg2-binary = "2.9.10"
python-dotenv = "1.1.0"
numpy = "^2.3.1"
openai = "^1.93.3"
langchain-openai = "^0.3.27"
psycopg2 = "^2.9.10"
langsmith = "^0.4.5"
openevals = "^0.1.0"
langchain-core = "^0.3.68"
langchain-postgres = "^0.0.15"
pydantic = "^2.11.7"
spacy = "^3.8.7"
neo4j = "^5.28.1"
pandas = "^2.3.1"
gradio = "^5.38.0"
langfuse = "^3.2.1"
fastapi = "^0.116.1"


[tool.poetry.group.dev.dependencies]
langchain-experimental = "^0.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
